chore: lldap fixes

+++ lldap/setup.sh
@@ -7,7 +7,7 @@ cd "$DIR" || exit 1
 
 source ../deploy/functions.sh
 
-if [[ ! -f .env ]]; then
+if [[ ! -f .env ]] && [[ -f env-example ]]; then
   cp env-example .env
 fi
